+++ lib/comment.py
@@ -114,7 +114,7 @@ def get_comment_regions(
 def allow_emmet_comments(view: sublime.View) -> bool:
     "Check if Emmet's Toggle Comment action can be applied at current view"
     if emmet.get_settings("toggle_comment"):
-        selectors = emmet.get_settings("comment_scopes", [])
+        selectors = syntax.selector_list("comment_scopes")
         caret = get_caret(view)
         return syntax.matches_selector(view, caret, selectors)